deflateAsync() for BlockCompressedOutputStream in case of SamSort
